Move SourceParsers to separate namespace
