+++ src/NullDev/Nemesis/SourceParser/PhpParserSourceParser.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace NullDev\Nemesis;
+namespace NullDev\Nemesis\SourceParser;
 
 use Exception;
+use NullDev\Nemesis\SourceParser;
 use NullDev\Skeleton\Definition\PHP\Methods\ConstructorMethod;
 use NullDev\Skeleton\Definition\PHP\Methods\GenericMethod;
 use NullDev\Skeleton\Definition\PHP\Methods\GetterMethod;
